LayoutTests/platform/glib/imported/w3c/web-platform-tests/geolocation-API/enabled-by-permission-policy-attribute-redirect-on-load.https.sub-expected.txt

CONSOLE MESSAGE: Permission policy 'Geolocation' check failed for document with origin 'https://not-web-platform.test:9443'.


PASS Permissions-Policy allow="geolocation" allows same-origin relocation
FAIL Permissions-Policy allow="geolocation" allows cross-origin relocation assert_true: geolocation expected true got false
